requirements.txt
# Function dependencies, for example:
# package>=version
pymysql==1.0.2
python-decouple==3.4
sqlalchemy==1.3.23
google-cloud-storage==1.36.2
google-cloud-logging==2.2.0
google-cloud-secret-manager==2.3.0
pysftp==0.2.9
paramiko==2.7.2